Updated System Manager
Cleaned up code for System Manager
