+++ Exomata/PlayerTransform.json
@@ -1,1 +1,1 @@
-{"posX":-0.10000000149011612,"posY":0.10000000149011612,"scaleX":0.10000000149011612,"scaleY":0.10000000149011612}
+{"posX":0.09999997168779373,"posY":0.10000000149011612,"scaleX":0.10000000149011612,"scaleY":0.10000000149011612}
